Refactor milestoning code
